tests/wpt/tests/css/css-values/tree-counting/sibling-index-keyframe-scale-dynamic.html

<!DOCTYPE html>
<title>CSS Values and Units Test: sibling-index() changing scale during @keyframes animation</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#tree-counting">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes --anim {
    from {
      scale: 0.7 calc(0.2 * sibling-index());
    }
    to {
      scale: 0.3 2;
    }
  }
  #target {
    animation: --anim 1000s step-end;
  }
</style>
<div>
  <div id="rm"></div>
  <div></div>
  <div id="target"></div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(target).scale, "0.7 0.6");
  }, "Initially, the sibling-index() is 3 for #target");

  test(() => {
    rm.remove();
    assert_equals(getComputedStyle(target).scale, "0.7 0.4");
  }, "Removing a preceding sibling of #target reduces the sibling-index()");

</script>
