LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-scale-dynamic-expected.txt


FAIL Initially, the sibling-index() is 3 for #target assert_equals: expected "0.7 0.6" but got "1"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "0.7 0.4" but got "1"
